+++ changelog.md
@@ -1,10 +1,13 @@
+# 1.0.6
+- Rename of mod under hood
+- Logic cleanup
+
 # 1.0.5
 - Added flags for Custom Position
 - Overhaul of the settings menu
 
-
 # 1.0.2
-- Added more option to preset locations:
+- Added more options to preset locations:
 	- Bottom Left
 	- Bottom Middle
 	- Bottom Right
@@ -12,7 +15,6 @@
 # 1.0.1
 - New setting! Popup Duration
 	- Added ability to choose fadeout animation speed
-
 
 # 1.0.0
 - Initial release.